🍧🍴 Checkpoint ./src/presenters/pages/team.js:271885/1331 ./src/components/team-profile/index.js:271885/7370 ./src/presenters/pages/team.styl:271885/71 ./src/components/team-profile/styles.styl:271885/104
