+++ src/components/team-profile/styles.styl
@@ -1,0 +1,5 @@
+@import  "../../components/constants"
+
+.teamUrl
+  color: text-on-secondary-background
+  font-weight: 600
