+++ src/presenters/pages/team.styl
@@ -38,10 +38,6 @@
 .teamMarketing
   border-top: 1px solid section-line
   padding-top: 1rem
-    
-.teamUrl
-  color: text-on-secondary-background
-  font-weight: 600
 
 .beta
   display: inline-flex
